requirements.txt
numpy>=1.23
azure-ai-documentintelligence
azure-ai-vision-imageanalysis
azure-core
dotenv
gradio
pytesseract
pdfplumber
reportlab
pillow
openai